pms selectors: Change selectors for `getRecentConversations`.
The parent commits introduced a new key
`recent_private_conversations` which fetches the private
conversations from the most recent 1000 private messages. This
commit changes the aforementioned selector to use the
`getRecentPrivateConversations` direct selector also introduced in
the same parent commit.

The effect of this is that the `PmConversationsCard` now displays
all the private conversations that a person might be participating
in instead of only from the last 100 messages (which was the amount
of messages that the previous iteration of the code fetched in
`doInitialFetch`, which directly determined the messages which were
scanned -- and how many conversations were returned).

`ownEmail` was removed in favor of `getOwnUser` in order to enable
the usage of `currentUser.user_id`.

The tests for `pmConversationsSelectors` has also been changed to
reflect the new data used and the way it is used. For example: we no
longer need to use the timestamps in order to sort conversations by
most recent because `max_message_id` satisifes that desire by
itself, and we don't need the returned conversations to have the
current user's ID in every `huddle` type conversation.