Add `recent_private_conversations` key in `register` endpoint.
The server added in zulip/zulip#11944 basic support for fetching the
most recent private conversations that a user was involved in.  The
endpoint returns the private conversations a user has had within the
past 1000 private messages they have sent/recieved.  The endpoint
returns data in the form of `[{user_ids, max_message_id}]` where
user_ids lists all the people involved in the conversation, and
max_message_id returns the message_id of the most recent message in
the conversation.

This commit introduces the reducers required to introduce this data
into the Redux state, during `doInitialFetch`.  It only catches the
`REALM_INIT` action at the moment, and the rest of the actions will
be added in a later commit with details.