redux: Store `state.narrows` as an Immutable.Map, not object-as-map.
An instance of #3949 and #3950.

When I wrote the commit message for e834f335d, I was most interested
in the exciting new ability to serialize/deserialize custom,
hand-made data types: in particular, ZulipVersion instances.

An important thing that *also* happened in that commit was the new
ability to store data structures from Immutable in Redux, without
any additional setup! `remotedev-serialize` comes with off-the-shelf
support for this. I gave this fact a brief shout-out near the end of
the commit message for a4c29e979.

So, do this, for the first time, paying careful attention to the
instructions in a4c29e979 for writing a migration (copied here, as
that commit message is quite long):

"""
    For making a value "live", where it wasn't before, the migration
    needs to:

    1) As input, take the previous shape of the data. Don't confuse this
    with the *current* way of storing the "dead" shape. Just like any
    other migration, the previous shape is the input.

    2) As output, give the "live" form of the data. Once it's in Redux,
    the replacer will take care of persisting it in the correct "dead"
    form.
"""

The previous shape of the data was an object-as-map, which
`Immutable.Map` takes as input quite cheerfully, and gives us our
"live" Immutable.Map instance.