[WIP] libdefs: Compose and use libdef for `react-native-webview` at 7.x.x.
The latest version FlowTyped has a libdef for is 6, unfortunately.

Ah, well. This is a best effort at replicating the types at
`react-native-community/react-native-webview@c4001338c`, tagged as
v7.6.0, the latest 7.x.

An improvement over the FlowTyped libdef is that WebView's props are
now exact. So, e.g., people will be warned not to use the
`useWebKit` prop, which was removed in 7.0.0.

See the entry in `docs/howto/libdefs.md` added in this commit for
more detail.