prettier: Set `arrowParens` explicitly to 'avoid'.
Done to follow the template-app change in
facebook/react-native@635ac1ba5, on the path to the RN v0.64
upgrade.

The default behavior on our Prettier version, according to the
doc [1], is already `arrowParens: 'avoid'`, but that default will
change to 'always' in Prettier v2.0.0. The RN commit sets it to
'avoid', explicitly, to avoid a silent change of behavior when
upgrading.

Might as well do the same, although the path to us actually using
Prettier v2 might be complicated; see
prettier/prettier-eslint-cli#304.

[1] https://prettier.io/docs/en/options.html#arrow-function-parentheses

That commit reports that the default behavior before Prettier 2.0 is
like setting `arrowParens: 'avoid'`. It's looking unlikely that we
have an easy path to using Prettier 2.0 (see
prettier/prettier-eslint-cli#304), but we might as well follow the
RN template app.