nav type tests [nfc]: Remove unused import.
I noticed this while looking at
  https://lgtm.com/projects/g/zulip/zulip-mobile/.

It looks like LGTM (and GitHub's "CodeQL", which uses it) isn't
ready to support Flow yet [1], and I think that's causing a lot of
noise into the alerts there. So, while this tool isn't quite ready
for us to use in CI yet, maybe we can check back on it at some
point? :)

[1] https://github.com/github/codeql/issues/6016