ios build: Follow a change in arg passed to `use_react_native!` in Podfile.
Done to follow the template-app change in
facebook/react-native@e599d6c5d, on the path to the RN v0.64
upgrade.

This isn't super well explained, but it looks like their claim for
this being needed starts by mentioning
@react-native-community/cli@e949e234b, which was released in v4.10.1
of @react-native-community/cli.

And `yarn why` tells me that we're on v4.14.0 of that package, so it
seems fine to take this.