createPersistor: Stop blending snapshots when we initiate writes.
This has the downside of ignoring state updates that happen while a
write is in progress. This isn't great, but it's better than not
persisting anything at all, or persisting data that doesn't
correspond to a value that an entire in-memory state object had at
some point.

We'll fix that in the upcoming commits, but in the meantime, this
makes the code much easier to mold into a reasonably performant
algorithm that uses `storage.multiSet` instead of several
`storage.setItem`s.