message list: Quickly handle safe areas in the message list.
This isn't ideal; it just puts the whole message list in a box
that's contained within the safe area. The message list was designed
to span the entire width of the screen, and some things look pretty
bad when the insets are nonzero [1]: the timestamp pills, for one
example, look like they get cut off artificially instead of being
tucked away somewhere offscreen. Things like that would best be
solved with a more in-depth design discussion.

A more complete solution might start with the "env" CSS function
[2], or, less optimally, a new WebView inbound event to be fired
whenever the insets change.

At least this solution lets things be visible and interactable that
weren't before.

I think this completes #3066 (modulo this message-list issue
discussed here), which is our current audit for safe-area handling.

[1] So far, I think this is just landscape mode on newer iOS
    devices, but I might be missing something.

[2] https://webkit.org/blog/7929/designing-websites-for-iphone-x/

Fixes: #3066