layout: Handle safe areas for screen-spanning "row" elements.
Each of these UI elements is a row that has meaningful content that
we need to keep within the safe areas, but the rest of the row (its
distinct background color, its touchable area, etc.) is meant to
extend through the insets to the extreme left and right of the
screen. See example screenshots at
  https://github.com/zulip/zulip-mobile/pull/4683#issue-619662894.

So, make that happen. We do so by giving the rows left and right
padding. This is easy and makes the row elements' styles pretty
intuitive, but it does mean we'll have to take care, in the next few
commits, not to add any padding to the elements that contain the
rows -- we don't want to double up the padding by mistake.

(This requirement ends up being kind of annoying on screens that
have a mix of these kinds of rows and regular elements. But this is
unfortunately the workable design we've found; see our architecture
doc on handling safe areas.)

Related: #3066