layout: Handle safe areas on some screens where it's a little tricky.
Tricky because the contents of these screens are mixed between
screen-spanning "row" elements (see a recent commit for what those
are) and regular elements.

These are all examples of "adjusting the wrapper's other children as
necessary to compensate", per our rules for handling safe areas in
the new doc added at the start of this series:

> If a React component is meant to occupy any insets, then its
> layout parent should adapt to allow it to do so, so that the inset
> distance isn't covered twice. This generally means omitting a
> padding `SafeAreaView` wrapper from the parent, if present,
> putting one in the child, and adjusting the wrapper's other
> children as necessary to compensate (e.g., by giving them
> `SafeAreaView`s, preferably with margin instead of padding, to
> show they don't care what fills the space).

Each component touched in this commit is a "parent" in that sense,
and has a jsdoc explaining which of its descendents make it
necessary for the component to occupy insets.