deps: Follow @react-native-{community -> async-storage}/async-storage move.
This seems to be a move / rename, not something more disruptive like
a rewrite. In fact, I don't expect this to be any more disruptive
than taking the next patch version in the library's natural
progression. If I add the new-location package at its first
available version, 1.13.0, and I keep the old one at its latest
available (and the one we're on currently), 1.12.1, I see very few
differences between them:

git diff --no-index \
  node_modules/@react-native-{community,async-storage}/async-storage

The package in its new location keeps the changelog entries from
when it was maintained at the old location. That changelog confirms
that 1.13.0 comes right after 1.12.1, and the 1.13.0 entry looks
like a match for what we saw in the `git diff --no-index` output.

So this shouldn't be disruptive at all, unless 1.13.0 has
unannounced breaking changes.

[1] For the output, see <insert link here>