Popup: Handle left and right insets.
In the next commit, note the TODOs in `EmojiAutocomplete`,
`PeopleAutocomplete`, and `StreamAutocomplete` (`TopicAutocomplete
isn't affected) for not double-handling these insets, which we do by
inappropriately padding the item rows.

The padding comes from some reusable components whose other
callsites *do* need the padding.

As the TODOs say, we should probably stop reusing those components,
and make new ones. The use-cases are different enough that this
probably makes sense to do in principle anyway. But another reason
not to reuse them is that it'd just be hard to deal with this issue
with the insets. `SafeAreaView` doesn't make it easy to
conditionally handle no edges or some edges; see
th3rdwave/react-native-safe-area-context#204.