api types: Add write-only `avatar_version` to User type
The other two properties in the comment don't appear in `realm_user`
items in the /register response, according to the doc. So, stop
mentioning them. (They do appear at toplevel in the /register
response, if `realm_user` is present in `fetch_event_types`; there,
they presumably apply to the logged-in user.)