wip android: Add "noNotifications" product flavor
This new product flavor supports building the app without
notification support. So in particular, we recommend it for
non-Kandra builds that want to do that, such as F-Droid's.

In the new product flavor, the app's internal notification API is
stubbed out, and the notifications feature is excluded gracefully.
See android/app/src/noNotifications/kotlin/notifications/ .

The Gradle task to build "release"-type APKs with the new flavor is
:app:assembleNoNotificationsRelease.

Fixes: # 5693